+++ weblate/settings_example.py
@@ -700,6 +700,7 @@ CRISPY_TEMPLATE_PACK = "bootstrap3"
 #     "weblate.checks.source.OptionalPluralCheck",
 #     "weblate.checks.source.EllipsisCheck",
 #     "weblate.checks.source.MultipleFailingCheck",
+#     "weblate.checks.format.MultipleUnnamedFormatsCheck",
 # )
 
 # List of automatic fixups