+++ docs/changes.rst
@@ -30,6 +30,7 @@ Not yet released.
 * Improved language code parsing.
 * Show translations in user language first in the list.
 * Renamed shapings to more generic name variants.
+* Added new quality checks: :ref:`check-unnamed-format`.
 
 Weblate 4.0.4
 -------------